git-svn-id: http://unused-160-33.aus.redhat.com/svn/penrose-studio/trunk@353 e99f1159-c40f-0410-9152-f9e4532946b1

+++ src/java/org/safehaus/penrose/studio/federation/nis/editor/NISPartitionsPage.java
@@ -191,18 +191,27 @@ public class NISPartitionsPage extends FormPage {
                         NISDomain repository = (NISDomain)ti.getData();
 
                         PartitionConfig partitionConfig = nisFederation.getPartitionConfig(repository.getName());
-                        PartitionConfig nssPartitionConfig = nisFederation.getPartitionConfig(repository.getName());
+                        if (partitionConfig != null) {
+                            nisFederation.removePartitionConfig(partitionConfig.getName());
+                            penroseClient.stopPartition(partitionConfig.getName());
+                            project.removeDirectory("partitions/"+partitionConfig.getName());
+                        }
+
+                        PartitionConfig ypPartitionConfig = nisFederation.getPartitionConfig(repository.getName()+"_"+NISFederation.YP);
+                        if (ypPartitionConfig != null) {
+                            nisFederation.removePartitionConfig(ypPartitionConfig.getName());
+                            penroseClient.stopPartition(ypPartitionConfig.getName());
+                            project.removeDirectory("partitions/"+ypPartitionConfig.getName());
+                        }
+
+                        PartitionConfig nssPartitionConfig = nisFederation.getPartitionConfig(repository.getName()+"_"+NISFederation.NSS);
+                        if (nssPartitionConfig != null) {
+                            nisFederation.removePartitionConfig(nssPartitionConfig.getName());
+                            penroseClient.stopPartition(nssPartitionConfig.getName());
+                            project.removeDirectory("partitions/"+nssPartitionConfig.getName());
+                        }
 
                         nisFederation.removePartition(repository);
-
-                        nisFederation.removePartitionConfig(partitionConfig.getName());
-                        penroseClient.stopPartition(partitionConfig.getName());
-
-                        nisFederation.removePartitionConfig(nssPartitionConfig.getName());
-                        penroseClient.stopPartition(nssPartitionConfig.getName());
-
-                        project.removeDirectory("partitions/"+partitionConfig.getName());
-                        project.removeDirectory("partitions/"+nssPartitionConfig.getName());
                     }
 
                     PenroseStudio penroseStudio = PenroseStudio.getInstance();
